Filter dead pixels only without changing the rest of the image
Previously a median filter of size 2 was used to replace single dead 
pixels with only minor changes to the rest of the image. Now a median 
filter of size 3 is used to find small dead clusters, too. By only 
updating the pixels that changed significantly, the rest of the image 
remains intact.